Force sign-out if refresh session failed
Also shows loading spinner when initializing authentication
