+++ src/components/Navigator.tsx
@@ -116,7 +116,9 @@ const Navigator: React.FC<Props> = (props) => {
     initialValueForNewTeamName,
   );
 
-  const { data: teams, refetch: refetchTeams, isFetching } = useGetTeamsQuery();
+  const { data: teams, refetch: refetchTeams, isFetching } = useGetTeamsQuery(undefined, {
+    skip: !userSub,
+  });
   const [selectingTeamId, setSelectingTeamId] = useState<string | null>(null);
   const { selectedTeam, refetch: refetchTeam, isFetching: isTeamFetching } = useSelectedTeam();
   const teamAvatar = useImageFromURL(
