+++ src/hooks/session.ts
@@ -1,21 +1,32 @@
 import * as CognitoIdentity from 'amazon-cognito-identity-js';
 import { useEffect, useState } from 'react';
-import { getSession, refreshSession } from '../auth';
+import { getSession, refreshSession, signout } from '../auth';
 
 export const useSession = () => {
   const [session, setSession] = useState<null | CognitoIdentity.CognitoUserSession >(null);
   const [isReady, setIsReady] = useState(false);
   useEffect(() => {
     (async () => {
-      let session = await getSession();
-      if (session === null) {
+      try {
+        let session = await getSession();
+        if (session === null) {
+          setIsReady(true);
+          return;
+        } else if (!session.isValid()) {
+          session = await refreshSession(session);
+        }
+        setSession(session);
         setIsReady(true);
-        return;
-      } else if (!session.isValid()) {
-        session = await refreshSession(session);
+      } catch (err: any) {
+        if (err.code === 'NotAuthorizedException') {
+          // something went wrong, sign out
+          // likely that the refresh token expired
+          await signout();
+          document.location.reload();
+        } else {
+          throw err;
+        }
       }
-      setSession(session);
-      setIsReady(true);
     })();
   }, []);
